Hopefully prevented crashing on race condition
Condition: If a user joined at the exact time a message was sent, it
will try to save the database for points as well as add the new user to
the database. This creates a race condition where the database gets a
user added while it's being enumerated causuing a crash. Locking seems
more graceful than try-catching